+++ packages/dev-utils/src/lib/logger.ts
@@ -1,7 +1,15 @@
-type logFunction = (...message: unknown[]) => void
+import ansis from 'ansis'
+
+type logFunction = (message?: string) => void
 
 export type Logger = {
   error: logFunction
   log: logFunction
   warn: logFunction
 }
+
+export const netlifyCommand = ansis.cyanBright
+
+export const netlifyCyan = ansis.rgb(40, 180, 170)
+
+export const netlifyBanner = netlifyCyan('⬥ Netlify')